requirements.txt
# ── Core Data ──────────────────────────────────────────────────
pandas==2.2.2
numpy==1.26.4
pyarrow==16.1.0

# ── Machine Learning (inference only) ──────────────────────────
scikit-learn==1.5.0
xgboost==2.0.3

# ── Dashboard ──────────────────────────────────────────────────
streamlit==1.35.0
plotly==5.22.0

# ── AI Querying ────────────────────────────────────────────────
anthropic==0.28.0

# ── Utilities ──────────────────────────────────────────────────
python-dotenv==1.0.1
pycountry==24.6.1
